Refactor customer service imports and update tests
 - Update customer controller imports to use specific service files
 - Remove customer.spec.ts and customer.ts as services are split into individual files
 - Add create.spec.ts, get.spec.ts, is-business.ts, status.ts, and update.spec.ts for customer services
 - Update user service tests to use createUser helper function
