+++ src/library/jest/helpers/user.ts
@@ -1,6 +1,6 @@
 import { faker } from "@faker-js/faker";
-import { CustomerServiceCreate } from "~/functions/customer/services/customer";
-import { Professions, User } from "~/functions/user";
+
+import { Professions, User, UserModel } from "~/functions/user";
 
 export const DEFAULT_GROUP = "all";
 
@@ -44,5 +44,9 @@ export const createUser = (
   filter: Partial<Pick<User, "customerId">> = {},
   props: Partial<User> = {}
 ) => {
-  return CustomerServiceCreate(getUserObject({ ...props, ...filter }));
+  const user = new UserModel({
+    ...getUserObject({ ...props, ...filter }),
+    isBusiness: true,
+  });
+  return user.save();
 };
